Layout base component #118 Layout selector + usage in base part #119
- renamed TypesRegistry.get/addContent to get/addContentType
- joined types definitions to one file typesSelector.ts
- removed region parsing functions from SingleComponent and passed region in meta
- moved types definitions to TypesRegistry.ts
