--- src/customXp/layoutSelector.ts
@@ -1,16 +1,0 @@
-import {TypesRegistry} from '../xpAdapter/TypesRegistry';
-import DefaultLayoutView from './layouts/_Layout';
-import ThreeColumnLayoutView, {THREE_COL_LAYOUT_NAME} from './layouts/ThreeColumnLayout';
-import CenteredLayoutView, {CENTERED_LAYOUT_NAME} from './layouts/CenteredLayout';
-
-TypesRegistry.addLayout(THREE_COL_LAYOUT_NAME, {
-    view: ThreeColumnLayoutView
-});
-
-TypesRegistry.addLayout(CENTERED_LAYOUT_NAME, {
-    view: CenteredLayoutView
-});
-
-TypesRegistry.addLayout("*", {
-    view: DefaultLayoutView
-});
