+++ src/pages/_app.tsx
@@ -7,11 +7,6 @@ import React from 'react';
 import Head from "next/head";
 import Layout from "../components/blocks/Layout";
 
-import "../customXp/contentSelector";
-import "../customXp/componentSelector";
-import "../customXp/partSelector";
-import "../customXp/layoutSelector";
-
 
 /**
  * Wraps all rendered components
