+++ src/xpAdapter/views/_SingleComponent.tsx
@@ -2,21 +2,14 @@
 
 import BaseComponent from "./_BaseComponent";
 import React from "react";
-import {FetchContentResult, getParentRegion, parseComponentPath} from "../guillotine/fetchContent";
+import {FetchContentResult} from "../guillotine/fetchContent";
 
 const SingleComponent = ({page, meta, content}: FetchContentResult) => {
-    if (!page?.regions || !Object.keys(page.regions)) {
-        // TODO: Fallback when view.regions is missing/empty (throw an error? 404?)
-    }
-
-    const cmpPath = parseComponentPath(meta.requestedComponent!);
-    const region = getParentRegion(page?.regions!, cmpPath);
-
-    if (!region) {
+    if (!meta.parentRegion) {
         // TODO: Handle missing target region (Throw a 404, "component path not found"?)
     }
 
-    const targetCompData = region.components.find(comp => comp.path === meta.requestedComponent);
+    const targetCompData = meta.parentRegion!.components.find(comp => comp.path === meta.requestedComponent);
 
     if (!targetCompData) {
         // TODO: Fallback for when target region or component is not found (render a placeholder? Throw a 404, "component path not found"?)
